admin can see private projects

+++ app/observers/project_observer.rb
@@ -6,7 +6,7 @@ class ProjectObserver < ActiveRecord::Observer
   def after_destroy record
     Broadcast.where(context_model_id: record.id, context_model_type: 'Project').destroy_all
     update_counters record, [:projects, :live_projects]
-    record.team.destroy
+    record.team.destroy if record.team
   end
 
   def after_update record
